chore: rename changes and commit to diff
