+++ internal/db/diff/templates/reset.sh
